Break up Help Menu & Replace About Menu Item with link to license (#7241)
* Move 'HelpMenu' to its own package

* Split up HelpMenu into package-protected classes per menu-item

* Drop AboutMenu

- Engine and map information are displayed in other locations,
  this makes most of the information in About Menu redundant
- The about menu duplicates the license text, that is already
  found in other locations. We can link to it instead
- The 'license' link in the about menu is not useful, it can
  neither be clicked nor copy-pasted
- The menu has an incorrect title, should be "about TripleA"
  rather than "about <Game>".
- It is a point of maintenance, the accreditation is out of
  date.

Beyond the above, it the about menu does not contain very
much information. The website link is perhaps the only
thing beyond this..

* Minor finishing fixes and formatting.

* Inline bug report menu and add menu item to open the game license

* Formatting

* Mark UnitStatsTable as @UtilityClass

* Remove unnecessary extends JMenu
