+++ game-core/src/main/java/games/strategy/triplea/ui/TripleAFrame.java
@@ -31,7 +31,6 @@ import games.strategy.engine.framework.IGame;
 import games.strategy.engine.framework.LocalPlayers;
 import games.strategy.engine.framework.ServerGame;
 import games.strategy.engine.framework.startup.ui.InGameLobbyWatcherWrapper;
-import games.strategy.engine.framework.system.SystemProperties;
 import games.strategy.engine.framework.ui.SaveGameFileChooser;
 import games.strategy.engine.history.HistoryNode;
 import games.strategy.engine.history.Round;
@@ -820,11 +819,6 @@ public final class TripleAFrame extends JFrame implements QuitHandler {
       ((ClientGame) game).shutDown();
       // an ugly hack, we need a better way to get the main frame
       new Thread(GameRunner::clientLeftGame).start();
-    }
-    if (SystemProperties.isMac()) {
-      // When leaving a game, reset the about handler to the default one, rather
-      // than the map-specific one set by HelpMenu.
-      MacOsIntegration.clearAboutHandler();
     }
   }
